requirements.txt
raglite[pandoc,ragas]>=0.6.2
streamlit>=1.32.0
python-dotenv>=1.0.0
pytest>=8.0.0 